src/main/webapp/app/entities/danh-muc-don-vi/delete/danh-muc-don-vi-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DanhMucDonViService } from '../service/danh-muc-don-vi.service';

import { DanhMucDonViDeleteDialogComponent } from './danh-muc-don-vi-delete-dialog.component';

describe('DanhMucDonVi Management Delete Component', () => {
  let comp: DanhMucDonViDeleteDialogComponent;
  let fixture: ComponentFixture<DanhMucDonViDeleteDialogComponent>;
  let service: DanhMucDonViService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [DanhMucDonViDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(DanhMucDonViDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DanhMucDonViDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DanhMucDonViService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
